Move deploy LCMS and PIX UI to run
