+++ common/services/slack/commands.js
@@ -2,11 +2,7 @@ const { releaseAndDeployPixBotTest } = require('../../../lib/services/releases')
 const releasesService = require('../../../lib/services/releases');
 const githubServices = require('../github');
 const axios = require('axios');
-const postSlackMessage = require('./surfaces/messages/post-message');
 
-const PIX_UI_REPO_NAME = 'pix-ui';
-const PIX_LCMS_REPO_NAME = 'pix-editor';
-const PIX_LCMS_APP_NAME = 'pix-lcms';
 const PIX_DATAWAREHOUSE_REPO_NAME = 'pix-db-replication';
 const PIX_DATAWAREHOUSE_APPS_NAME = ['pix-datawarehouse', 'pix-datawarehouse-ex'];
 
@@ -29,10 +25,6 @@ function getErrorAppMessage(appName) {
   return `Erreur lors du déploiement de ${appName} en production.`;
 }
 
-function getErrorReleaseMessage(release, appName) {
-  return `Erreur lors du déploiement de la release '${release}' pour ${appName} en production.`;
-}
-
 function _isReleaseTypeInvalid(releaseType) {
   return !['major', 'minor', 'patch'].includes(releaseType);
 }
@@ -53,32 +45,7 @@ async function publishAndDeployRelease(repoName, appNamesList = [], releaseType,
   }
 }
 
-async function publishAndDeployPixUI(repoName, releaseType, responseUrl) {
-  if (_isReleaseTypeInvalid(releaseType)) {
-    releaseType = 'minor';
-  }
-  const releaseTagBeforeRelease = await githubServices.getLatestReleaseTag(repoName);
-  await releasesService.publishPixRepo(repoName, releaseType);
-  const releaseTagAfterRelease = await githubServices.getLatestReleaseTag(repoName);
-  await releasesService.deployPixUI();
-
-  if (releaseTagBeforeRelease === releaseTagAfterRelease) {
-    postSlackMessage(getErrorReleaseMessage(releaseTagAfterRelease, repoName));
-  } else {
-    sendResponse(responseUrl, getSuccessMessage(releaseTagAfterRelease, repoName));
-    postSlackMessage(`[PIX-UI] New version deployed : (${releaseTagAfterRelease})`);
-  }
-}
-
 module.exports = {
-
-  async createAndDeployPixLCMS(payload) {
-    await publishAndDeployRelease(PIX_LCMS_REPO_NAME, [PIX_LCMS_APP_NAME], payload.text, payload.response_url);
-  },
-
-  async createAndDeployPixUI(payload) {
-    await publishAndDeployPixUI(PIX_UI_REPO_NAME, payload.text, payload.response_url);
-  },
 
   async createAndDeployPixDatawarehouse(payload) {
     await publishAndDeployRelease(PIX_DATAWAREHOUSE_REPO_NAME, PIX_DATAWAREHOUSE_APPS_NAME, payload.text, payload.response_url);
