+++ lib/services/releases.js
@@ -5,7 +5,6 @@ const ScalingoClient = require('../../common/services/scalingo-client');
 const exec = util.promisify(require('child_process').exec);
 
 const RELEASE_PIX_SCRIPT = 'release-pix-repo.sh';
-const DEPLOY_PIX_UI_SCRIPT = 'deploy-pix-ui.sh';
 
 module.exports = {
 
@@ -44,11 +43,6 @@ module.exports = {
       console.error(err);
       throw err;
     }
-  },
-
-  async deployPixUI() {
-    const args = [config.github.owner, 'pix-ui'];
-    await _runScriptWithArgument(DEPLOY_PIX_UI_SCRIPT, ...args);
   },
 
   async releaseAndDeployPixBotTest(versionType) {
